Refactor to make remote builds work same as local
This fixes the problem of the remote build being able to report the correct
line numbers for source files with parse errors. But beyond that, it implements
https://github.com/whatwg/html-build/pull/28#issuecomment-139674977 by
making the local and server-side builds run in exactly the same way.

https://github.com/domenic/wattsi-server/pull/4 needs to be enabled on the
server side in order for this to work as expected for remote builds.