Keep entities stable and only generate them on demand
This fixes part of #62 by moving the generation of entity “tables”
outside of the main build script. Instead, we’ll run it only once all
browsers agree a change needs to be made.

build.sh doesn’t have a cache since we’ll run it very rarely.

This also moves the two entity files we had in whatwg/html here, to
keep everything about entities in a single place.

This requires a corresponding commit to whatwg/html.