Download a smaller list of bugs from W3C's Bugzilla
Limit the search to bugs that have a URL and are in interesting
components.

This shrinks the downloaded w3cbugs.cvs from 3211 bugs (242 kB) to 265
bugs (40 kB), with corresponding savings in transfer time both when
downloaded and then uploading to the wattsi build server.

This removes the following HTML Checker and SVG bugs from the output:
https://www.w3.org/Bugs/Public/show_bug.cgi?id=13648
https://www.w3.org/Bugs/Public/show_bug.cgi?id=13649
https://www.w3.org/Bugs/Public/show_bug.cgi?id=16125
https://www.w3.org/Bugs/Public/show_bug.cgi?id=23157
https://www.w3.org/Bugs/Public/show_bug.cgi?id=23165
https://www.w3.org/Bugs/Public/show_bug.cgi?id=23753

Based on findings from https://github.com/whatwg/html/issues/619