Deploy commit snapshots to directories, not files
This matches other WHATWG specs, but more importantly will make it
easier to compress the index.html on disk. With nginx, that requires
an empty index.html and a compressed index.html.gz, and putting them
in directories avoids both being listed in
https://html.spec.whatwg.org/commit-snapshots/, and the compressed
file size showing up as well.

This is a bit weird but WAI in nginx:
https://trac.nginx.org/nginx/ticket/1367