Stop buffering wattsi output when run locally
This change causes messages from wattsi to get written to the console in real
time as wattsi emits them.

We so far had essentially been buffering output from wattsi — writing the wattsi
output to file and then emitting that after wattsi exited.

The advantage of that was, it allowed for the same output handling as the case
where wattsi is run remotely. The disadvantage was that is prevented us from
watching the progress of wattsi in real time as it runs.

Because wattsi has run relatively quickly — taking on the order of just ~6
seconds from start to finish — the lack of ability to watch wattsi progress in
real time has so far not been a big problem. However, the addition of mechanism
for syntax-highlighting `pre` elements noticeably increases the wattsi
execution time, to the degree that not being able to see messages from wattsi in
real time can make it appear to be hanging.

So this change, along with the “Emit each HTML output filename as it gets written”
change in https://github.com/whatwg/wattsi/pull/63 provides a way to watch the
progress of wattsi in detail as it runs, so we can see exactly what is happening
and when (and know what steps wattsi is running are consuming the most time).