Switch from Apache to nginx for build.sh --docker
Fixes https://github.com/whatwg/html-build/issues/140.

Note that none of the redirects from misc-server are included here,
and weren't before either, in particular /multipage/images/ won't
redirect to /images/. To fix this well the configuration should be
shared, which is a bit of work, and so it's left broken for now.