Reland "Deploy commit snapshots to directories, not files"
This relands commit 33a5b3254e91e413dd9819460f47e9b6203a0eb3 with a
different approach which is more similar to the non-HTML
deploy.sh. First create a directory, then rsync that.

This was tested manually with a bogus commit using the live
server. --chmod=D755,F644 was added because the local permissions are
otherwise copied, and there was already some variation on the live
server, both 0644 and 0664 files. That was normalized.