Use HTML checker binary image, not system Java
This change makes the HTML-spec deploy build download and use the
(Linux) binary image of the HTML checker rather than downloading and
using the jar release of the checker and relying on using a system Java
runtime to run that jar.

Note that we call ./vnu-runtime-image/bin/java rather than the
./vnu-runtime-image/bin/vnu convenience wrapper script because we need
to feed java the -Xmx1g argument.