Include the default html source folder
While following the steps here, I've noticed the `./build.sh` - including the docker run - never asked me for the HTML source folder. The script does a very convenient lookup for a `../html` source folder instead.

This commit also adds a quick paragraph explaining how to set a custom location, based on the source code of the build script. This may be convenient to other users.

These are all suggestions to the documentation. I'd be happy to change the text if necessary, just let me know.