Add LICENSE file; CC Attribution 4.0 International
This change adds an explicit LICENSE file to the repo. (The same
Creative Commons Attribution 4.0 International License that we use for
most other WHATWG repos.)

Otherwise without this change, it’s unclear to others if we’re licensing
the code for them to use. And there are some files in particular such as
https://github.com/whatwg/html-build/blob/master/entities/out/entities.json
which have some general utility to others.

Closes #249.