Remove DockerHub caching of CI builds (for now)
We'd like to move this to GitHub actions, which has simpler and less manual options for caching Docker layers. For now, let's remove the caching that we're doing, so that the transition to GitHub actions is simpler. This will make CI builds and deploys of whatwg/html-build and whatwg/html slower, but we'll survive.