Re-do CI build
Per the plan in https://github.com/whatwg/meta/issues/173#issuecomment-697050032, this revamps the CI build to reduce coupling between the html-build and html repositories. Now, running CI (currently Travis) for the html-build repository produces a whatwg/html-build Docker container, pushed to Docker Hub, which can be given an input directory containing the contents of whatwg/html, and an output directory.

This will require corresponding changes on the whatwg/html side to make use of this new architecture.