Fix PDF generation
Subresources were not downloading correctly, likely due to the old version of Prince having out-of-date SSL support. (I noticed it emitting warnings about invalid certificates for resources.whatwg.org.)

This upgrades to the latest version of Prince. Doing so requires abandoning pdfsizeopt, as pdfsizeopt is not compatible with modern Prince. Oh well.

This allows us to use a .deb-based installation procedure for Prince instead of compiling from source, which is nice. In particular we no longer need to manually install dependencies.