Avoid generating review drafts for merge commits.
This avoids generating review drafts for merge commits unless the merge
commit has modifications to the review draft (which I believe means that
it's different from *all* parents of the merge).

The code prior to this change tries to generate a review draft if the
first parent of the merge commit introduces a review draft.  (This
happens if "git merge main" is done on a feature branch and main has
introduced a review draft.)