Pass --no-pager to git log command.
Without --no-pager, this command can (depending on the user's pager
settings) cause build.sh to show just a full screen git log in the
user's pager (for example, if PAGER="less -Cemf").  This behavior is
unexpected, and it seems like the intended behavior here is to have the
git log appear as a continuous part of the output.