Add initial variable processing
This supports the work in https://github.com/whatwg/html/pull/11392 with a new processor found in variables.rs, documented there.

This also updates the parser to store the line numbers for each element, for use in error messages.

For now this contains commented-out code to error on `<var>`s outside of `algorithm` and `var-scope` scopes, since CI will prevent us from merging a html-build change that cannot build the current main branch of whatwg/html. After the HTML PR is merged, we can uncomment that code.