Add experimental support to build using Bikeshed
The new --bikeshed flag enables Bikeshed mode. The wattsi2bikeshed.js
script converts the source from Wattsi to Bikeshed syntax.

The hardest part of this is ensuring that cross-references still resolve
to the same thing. Wattsi's <span> and <i> for autolinks are converted
to <a>, but changes to HTML itself will likely be needed for some cases
where the data-x="" (with hyphens) is unambiguous but where Bikeshed's
linking text (with spaces) is ambiguous and would require a for=""
attribute.

What doesn't work:

  - Some cross-references as per above.
  - Some manual changes to HTML are still needed:
    https://github.com/whatwg/html/tree/bikeshed-experiment
  - Existing IDs aren't always preserved.
  - Simplification of HTML into Bikeshed shorthands like {{Node}}.
  - subdfn, which has no equivalent in Bikeshed.
  - Variants like the dev edition aren't properly supported.